+++ test/integration/braintree/merchant_gateway_spec.js
@@ -16,7 +16,7 @@ describe("MerchantGateway", function () {
       return gateway.merchant.create(
         {
           email: "name@email.com",
-          countryCodeAlpha3: "USA",
+          countryCodeAlpha3: "GBR",
           paymentMethods: ["credit_card", "paypal"],
         },
         function (err, response) {
@@ -28,10 +28,10 @@ describe("MerchantGateway", function () {
           assert.isNotNull(merchant.id);
           assert.equal(merchant.email, "name@email.com");
           assert.equal(merchant.companyName, "name@email.com");
-          assert.equal(merchant.countryCodeAlpha3, "USA");
-          assert.equal(merchant.countryCodeAlpha2, "US");
-          assert.equal(merchant.countryCodeNumeric, "840");
-          assert.equal(merchant.countryName, "United States of America");
+          assert.equal(merchant.countryCodeAlpha3, "GBR");
+          assert.equal(merchant.countryCodeAlpha2, "GB");
+          assert.equal(merchant.countryCodeNumeric, "826");
+          assert.equal(merchant.countryName, "United Kingdom");
 
           let credentials = response.credentials;
 
@@ -75,14 +75,14 @@ describe("MerchantGateway", function () {
   describe("create_multi_currency", function () {
     it("creates a multi-currency merchant", function (done) {
       let gateway = new braintree.BraintreeGateway({
-        clientId: "client_id$development$signup_client_id",
-        clientSecret: "client_secret$development$signup_client_secret",
+        clientId: "client_id$development$integration_client_id",
+        clientSecret: "client_secret$development$integration_client_secret",
       });
 
       return gateway.merchant.create(
         {
           email: "name@email.com",
-          countryCodeAlpha3: "USA",
+          countryCodeAlpha3: "GBR",
           paymentMethods: ["credit_card", "paypal"],
           currencies: ["GBP", "USD"],
         },
@@ -95,10 +95,10 @@ describe("MerchantGateway", function () {
           assert.isNotNull(merchant.id);
           assert.equal(merchant.email, "name@email.com");
           assert.equal(merchant.companyName, "name@email.com");
-          assert.equal(merchant.countryCodeAlpha3, "USA");
-          assert.equal(merchant.countryCodeAlpha2, "US");
-          assert.equal(merchant.countryCodeNumeric, "840");
-          assert.equal(merchant.countryName, "United States of America");
+          assert.equal(merchant.countryCodeAlpha3, "GBR");
+          assert.equal(merchant.countryCodeAlpha2, "GB");
+          assert.equal(merchant.countryCodeNumeric, "826");
+          assert.equal(merchant.countryName, "United Kingdom");
 
           let credentials = response.credentials;
 
@@ -117,7 +117,7 @@ describe("MerchantGateway", function () {
           )[0];
 
           assert.isNotNull(usdMerchantAccount);
-          assert.equal(usdMerchantAccount.default, true);
+          assert.equal(usdMerchantAccount.default, false);
           assert.equal(usdMerchantAccount.currencyIsoCode, "USD");
 
           let gbpMerchantAccount = merchantAccounts.filter(
@@ -125,7 +125,7 @@ describe("MerchantGateway", function () {
           )[0];
 
           assert.isNotNull(gbpMerchantAccount);
-          assert.equal(gbpMerchantAccount.default, false);
+          assert.equal(gbpMerchantAccount.default, true);
           assert.equal(gbpMerchantAccount.currencyIsoCode, "GBP");
 
           done();
@@ -135,14 +135,14 @@ describe("MerchantGateway", function () {
 
     it("creates a paypal-only merchant", function (done) {
       let gateway = new braintree.BraintreeGateway({
-        clientId: "client_id$development$signup_client_id",
-        clientSecret: "client_secret$development$signup_client_secret",
+        clientId: "client_id$development$integration_client_id",
+        clientSecret: "client_secret$development$integration_client_secret",
       });
 
       return gateway.merchant.create(
         {
           email: "name@email.com",
-          countryCodeAlpha3: "USA",
+          countryCodeAlpha3: "GBR",
           paymentMethods: ["paypal"],
           currencies: ["GBP", "USD"],
           paypalAccount: {
@@ -159,10 +159,10 @@ describe("MerchantGateway", function () {
           assert.isNotNull(merchant.id);
           assert.equal(merchant.email, "name@email.com");
           assert.equal(merchant.companyName, "name@email.com");
-          assert.equal(merchant.countryCodeAlpha3, "USA");
-          assert.equal(merchant.countryCodeAlpha2, "US");
-          assert.equal(merchant.countryCodeNumeric, "840");
-          assert.equal(merchant.countryName, "United States of America");
+          assert.equal(merchant.countryCodeAlpha3, "GBR");
+          assert.equal(merchant.countryCodeAlpha2, "GB");
+          assert.equal(merchant.countryCodeNumeric, "826");
+          assert.equal(merchant.countryName, "United Kingdom");
 
           let credentials = response.credentials;
 
@@ -181,7 +181,7 @@ describe("MerchantGateway", function () {
           )[0];
 
           assert.isNotNull(usdMerchantAccount);
-          assert.equal(usdMerchantAccount.default, true);
+          assert.equal(usdMerchantAccount.default, false);
           assert.equal(usdMerchantAccount.currencyIsoCode, "USD");
 
           let gbpMerchantAccount = merchantAccounts.filter(
@@ -189,7 +189,7 @@ describe("MerchantGateway", function () {
           )[0];
 
           assert.isNotNull(gbpMerchantAccount);
-          assert.equal(gbpMerchantAccount.default, false);
+          assert.equal(gbpMerchantAccount.default, true);
           assert.equal(gbpMerchantAccount.currencyIsoCode, "GBP");
 
           done();
@@ -199,14 +199,14 @@ describe("MerchantGateway", function () {
 
     it("allows creation of non-US merchant if onboarding application is internal", function (done) {
       let gateway = new braintree.BraintreeGateway({
-        clientId: "client_id$development$signup_client_id",
-        clientSecret: "client_secret$development$signup_client_secret",
+        clientId: "client_id$development$integration_client_id",
+        clientSecret: "client_secret$development$integration_client_secret",
       });
 
       return gateway.merchant.create(
         {
           email: "name@email.com",
-          countryCodeAlpha3: "JPN",
+          countryCodeAlpha3: "GBR",
           paymentMethods: ["paypal"],
           paypalAccount: {
             clientId: "fake_client_id",
@@ -222,10 +222,10 @@ describe("MerchantGateway", function () {
           assert.isNotNull(merchant.id);
           assert.equal(merchant.email, "name@email.com");
           assert.equal(merchant.companyName, "name@email.com");
-          assert.equal(merchant.countryCodeAlpha3, "JPN");
-          assert.equal(merchant.countryCodeAlpha2, "JP");
-          assert.equal(merchant.countryCodeNumeric, "392");
-          assert.equal(merchant.countryName, "Japan");
+          assert.equal(merchant.countryCodeAlpha3, "GBR");
+          assert.equal(merchant.countryCodeAlpha2, "GB");
+          assert.equal(merchant.countryCodeNumeric, "826");
+          assert.equal(merchant.countryName, "United Kingdom");
 
           let credentials = response.credentials;
 
@@ -242,59 +242,7 @@ describe("MerchantGateway", function () {
           let merchantAccount = merchantAccounts[0];
 
           assert.equal(merchantAccount.default, true);
-          assert.equal(merchantAccount.currencyIsoCode, "JPY");
-
-          done();
-        }
-      );
-    });
-
-    it("defaults to USD for non-US merchant if onboarding application is internal and country currency not supported", function (done) {
-      let gateway = new braintree.BraintreeGateway({
-        clientId: "client_id$development$signup_client_id",
-        clientSecret: "client_secret$development$signup_client_secret",
-      });
-
-      return gateway.merchant.create(
-        {
-          email: "name@email.com",
-          countryCodeAlpha3: "YEM",
-          paymentMethods: ["paypal"],
-          paypalAccount: {
-            clientId: "fake_client_id",
-            clientSecret: "fake_client_secret",
-          },
-        },
-        function (err, response) {
-          assert.isNull(err);
-          assert.isTrue(response.success);
-
-          let merchant = response.merchant;
-
-          assert.isNotNull(merchant.id);
-          assert.equal(merchant.email, "name@email.com");
-          assert.equal(merchant.companyName, "name@email.com");
-          assert.equal(merchant.countryCodeAlpha3, "YEM");
-          assert.equal(merchant.countryCodeAlpha2, "YE");
-          assert.equal(merchant.countryCodeNumeric, "887");
-          assert.equal(merchant.countryName, "Yemen");
-
-          let credentials = response.credentials;
-
-          assert.isNotNull(credentials.accessToken);
-          assert.equal(credentials.accessToken.indexOf("access_token"), 0);
-          assert.isNotNull(credentials.refreshToken);
-          assert.isNotNull(credentials.expiresAt);
-          assert.equal(credentials.tokenType, "bearer");
-
-          let merchantAccounts = merchant.merchantAccounts;
-
-          assert.equal(merchantAccounts.length, 1);
-
-          let merchantAccount = merchantAccounts[0];
-
-          assert.equal(merchantAccount.default, true);
-          assert.equal(merchantAccount.currencyIsoCode, "USD");
+          assert.equal(merchantAccount.currencyIsoCode, "GBP");
 
           done();
         }
@@ -303,14 +251,14 @@ describe("MerchantGateway", function () {
 
     it("returns error if invalid currency is passed", function (done) {
       let gateway = new braintree.BraintreeGateway({
-        clientId: "client_id$development$signup_client_id",
-        clientSecret: "client_secret$development$signup_client_secret",
+        clientId: "client_id$development$integration_client_id",
+        clientSecret: "client_secret$development$integration_client_secret",
       });
 
       return gateway.merchant.create(
         {
           email: "name@email.com",
-          countryCodeAlpha3: "USA",
+          countryCodeAlpha3: "GBR",
           paymentMethods: ["paypal"],
           currencies: ["FAKE", "USD"],
           paypalAccount: {